Update sankey.py
Slight modifications that were helpful for me in producing Sankey plots for my work. 

As the "aspect" keyword didn't allow a specification of the figure size, I've added explicit arguments for figure height and width. 

I also added an argument for a plot directory, so that the user can specify where to save the figure if they want.

Feel free to integrate or ignore, as you prefer